Extracts Job interface to its own package
Pages have a Links message which returns an array of its child Pages. In
order to simpify testing the Worker the Job interface is used in place
of Page. The return type of the Links method in the Job interface is
[]Job. The return type for the Links method of a Struct implementing this
Interface has to be []Job. Seems best for that reason to extract it and
make it more obvious that it is a shared Interface